[#131] Fix and improve Receiver

+++ src/back/task/Receiver.js
@@ -73,7 +73,7 @@ class Receiver {
    */
   #itemWrittenBytes;
   /**
-   * Number of received items so far.
+   * Number of completely received, skipped, or errored items so far.
    * @type {number}
    */
   #numRecvItem;
@@ -211,14 +211,13 @@ class Receiver {
               // mark it failed, and go to next item.
               // TODO mark the item failed.
               try {
-                if (this.#itemHandle) {
+                if (this.#itemHandle && this.#itemName) {
                   await this.#itemHandle.close();
-                }
-                if (this.#itemName) {
                   await fs.rm(path.join(this.#recvPath, this.#itemName), {force: true});
                 }
               } finally {
                 this.#itemHandle = null;
+                this.#itemName = null;
                 this.#itemFlag = 'next';
                 this.#numRecvItem++;
                 this.sendHeader();
@@ -234,11 +233,16 @@ class Receiver {
           break;
 
         case 'new':
+          if (this.#itemHandle) {
+            // Close previous item handle.
+            await this.#itemHandle.close();
+            this.#itemHandle = null;
+            this.#numRecvItem++;
+          }
           this.#itemName = path.join(this.#recvHeader.dir, this.#recvHeader.name);
           if (this.#recvHeader.type === 'directory') {
             this.#haveParsedHeader = false;
             this.#itemSize = 0;
-            this.#numRecvItem++;
             try {
               await fs.mkdir(path.join(this.#recvPath, this.#itemName));
             } catch (err) {
@@ -251,15 +255,11 @@ class Receiver {
               }
             }
             this.#itemFlag = 'ok';
+            this.#numRecvItem++;
             this.sendHeader();
           }
           else if (this.#recvHeader.type === 'file') {
             try {
-              if (this.#itemHandle) {
-                this.#numRecvItem++;
-                // Close previous item handle.
-                await this.#itemHandle.close();
-              }
               this.#itemHandle = await fs.open(path.join(this.#recvPath, this.#itemName), 'wx');
             } catch {
               // File already exists.
@@ -267,6 +267,7 @@ class Receiver {
               // 1. Ignore the file.
               // 2. Create the file with another name.
               this.#itemHandle = null;
+              this.#itemName = null;
               this.#itemFlag = 'next';
               this.#numRecvItem++;
               this.sendHeader();
@@ -293,6 +294,8 @@ class Receiver {
         case 'end':
           this.#setState(STATE.OTHER_END);
           // Close previous item handle and delete.
+          // When deleting the file the handle should be truthy,
+          // to make sure the file is actually written by tiShare.
           if (this.#itemHandle && this.#itemName) {
             await this.#itemHandle.close();
             this.#itemHandle = null;
